@@ -31,104 +31,113 @@
     "e2e": "ng e2e",
     "docs": "compodoc -p src/tsconfig.app.json -d docs",
     "docs:serve": "compodoc -p src/tsconfig.app.json -d docs -s",
-    "release:changelog": "npm run conventional-changelog -- -p angular -i CHANGELOG.md -s"
+    "release:changelog": "npm run conventional-changelog -- -p angular -i CHANGELOG.md -s",
+    "postinstall": "ngcc"
   },
   "dependencies": {
-    "@agm/core": "^1.0.0-beta.5",
-    "@angular-devkit/build-angular": "^0.13.0",
-    "@angular-devkit/build-ng-packagr": "^0.803.21",
-    "@angular/animations": "^7.2.15",
-    "@angular/cdk": "^7.3.7",
-    "@angular/common": "^7.2.15",
-    "@angular/compiler": "^7.2.15",
-    "@angular/core": "^7.2.15",
-    "@angular/forms": "^7.2.15",
-    "@angular/http": "^7.2.15",
-    "@angular/material": "^7.3.7",
-    "@angular/platform-browser": "^7.2.15",
-    "@angular/platform-browser-dynamic": "^7.2.15",
-    "@angular/router": "^7.2.15",
+    "@agm/core": "^3.0.0-beta.0",
+    "@angular-devkit/build-angular": "~0.1001.1",
+    "@angular-devkit/build-ng-packagr": "~0.1001.1",
+    "@angular/animations": "^10.1.1",
+    "@angular/cdk": "^10.0.0",
+    "@angular/common": "^10.1.1",
+    "@angular/compiler": "^10.1.1",
+    "@angular/core": "^10.1.1",
+    "@angular/forms": "^10.1.1",
+    "@angular/localize": "^10.1.1",
+    "@angular/material": "^10.0.0",
+    "@angular/platform-browser": "^10.1.1",
+    "@angular/platform-browser-dynamic": "^10.1.1",
+    "@angular/router": "^10.1.1",
     "@google/markerclusterer": "^2.0.0",
-    "@nebular/auth": "3.0.0",
-    "@nebular/bootstrap": "3.0.0",
-    "@nebular/security": "3.0.0",
-    "@nebular/theme": "^5.1.0",
-    "@ng-bootstrap/ng-bootstrap": "^4.1.2",
-    "@progress/kendo-angular-dateinputs": "^3.7.3",
-    "@progress/kendo-angular-intl": "^1.4.0",
-    "@progress/kendo-angular-l10n": "^1.1.0",
+    "@nebular/auth": "^5.1.0",
+    "@nebular/bootstrap": "^5.1.0",
+    "@nebular/eva-icons": "^5.1.0",
+    "@nebular/security": "^6.0.0",
+    "@nebular/theme": "^6.0.0",
+    "@ng-bootstrap/ng-bootstrap": "^7.0.0",
+    "@progress/kendo-angular-common": "^1.2.2",
+    "@progress/kendo-angular-dateinputs": "^4.2.2",
+    "@progress/kendo-angular-intl": "^2.0.1",
+    "@progress/kendo-angular-l10n": "^2.0.1",
+    "@progress/kendo-angular-popup": "^3.0.5",
     "@progress/kendo-theme-default": "latest",
-    "angular-highcharts": "^8.0.3",
+    "@techiediaries/ngx-qrcode": "^9.1.0",
+    "@types/jquery": "^3.5.1",
+    "angular-csv-ext": "^1.0.3",
+    "angular-highcharts": "^10.0.0",
     "angular-resizable-element": "^3.2.4",
     "angular-tooltips": "^1.2.2",
     "angular2-chartjs": "^0.5.1",
     "angular2-draggable": "^2.2.4",
-    "angular5-csv": "^0.2.11",
     "bootstrap": "^4.3.1",
     "classlist.js": "1.1.20150312",
     "core-js": "2.5.1",
     "d3": "^5.16.0",
+    "eva-icons": "^1.1.3",
     "file-saver": "^2.0.2",
+    "font-awesome": "^4.7.0",
     "html2canvas": "^1.0.0-rc.1",
     "html2pdf.js": "^0.9.1",
     "intl": "1.2.5",
+    "iv-viewer": "^2.0.1",
+    "jquery": "^3.5.1",
     "jspdf": "^1.5.3",
     "jspdf-autotable": "^3.1.1",
     "localforage": "^1.8.1",
     "material-design-icons": "^3.0.1",
     "moment": "^2.24.0",
-    "ng-pick-datetime": "^7.0.0",
+    "nebular-icons": "^1.1.0",
+    "ng-packagr": "^10.1.0",
+    "ng-pick-datetime-ex": "^9.0.8",
     "ng2-completer": "^2.0.8",
-    "ng2-image-viewer": "^2.0.2",
     "ng2-pdf-viewer": "^5.3.4",
-    "ng2-smart-table": "^1.4.0",
     "ngx-image-cropper": "^1.3.9",
     "ngx-print": "1.2.0-beta.3",
-    "ngx-qrcode2": "0.0.9",
-    "ngx-qrcode3": "^0.3.0",
     "node-sass": "^4.14.1",
     "normalize.css": "6.0.0",
     "pace-js": "1.0.2",
     "roboto-fontface": "0.8.0",
-    "rxjs": "6.3.0",
+    "rxjs": "6.6.3",
     "rxjs-compat": "^6.0.0",
-    "sass-loader": "^9.0.2",
     "stream": "0.0.2",
     "timers": "^0.1.1",
     "topojson-client": "^3.1.0",
+    "tslib": "^2.0.0",
     "typeface-exo": "0.0.22",
-    "web-animations-js": "2.2.5",
+    "web-animations-js": "^2.3.2",
     "xlsx": "^0.14.3",
-    "zone.js": "^0.8.29"
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/core": "^10.0.4",
-    "@angular/cli": "^7.3.8",
-    "@angular/compiler-cli": "^7.2.15",
-    "@angular/language-service": "7.0.0",
+    "@angular-devkit/core": "^10.1.1",
+    "@angular/cli": "^10.1.1",
+    "@angular/compiler-cli": "^10.1.1",
+    "@angular/language-service": "10.1.1",
     "@compodoc/compodoc": "^1.1.10",
+    "@fortawesome/fontawesome-free": "^5.14.0",
     "@types/d3": "^5.7.2",
     "@types/jasmine": "2.5.54",
     "@types/jasminewd2": "2.0.3",
-    "@types/node": "6.0.90",
-    "codelyzer": "3.2.1",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^5.1.2",
     "conventional-changelog-cli": "1.3.4",
     "husky": "0.13.3",
-    "jasmine-core": "2.6.4",
-    "jasmine-spec-reporter": "4.1.1",
-    "karma": "^4.4.1",
-    "karma-chrome-launcher": "2.1.1",
+    "jasmine-core": "~3.5.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~5.0.0",
+    "karma-chrome-launcher": "~3.1.0",
     "karma-cli": "1.0.1",
-    "karma-coverage-istanbul-reporter": "1.3.0",
-    "karma-jasmine": "1.1.0",
-    "karma-jasmine-html-reporter": "0.2.2",
+    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
     "npm-run-all": "4.0.2",
-    "protractor": "^5.4.2",
+    "protractor": "~7.0.0",
     "rimraf": "2.6.1",
+    "sass-loader": "^7.1.0",
     "stylelint": "^9.10.1",
     "ts-node": "3.2.2",
-    "tslint": "5.7.0",
-    "tslint-language-service": "^0.9.9",
-    "typescript": "3.1.3"
+    "tslint": "~6.1.0",
+    "typescript": "4.0.2"
   }
 }
