@@ -7,19 +7,12 @@ import * as _ from 'lodash';
 import { CsvService } from '../../services/csv/csv.service';
 import { AccountService } from '../../services/account.service';
 import { UserService } from '../../services/user.service';
-import { VoucherdetailComponent } from '../../acounts-modals/voucherdetail/voucherdetail.component';
-import { OrderdetailComponent } from '../../acounts-modals/orderdetail/orderdetail.component';
-import { TripdetailComponent } from '../../acounts-modals/tripdetail/tripdetail.component';
+
 
 // import { Injectable } from '@angular/core';
-// import { HttpClient, HttpHeaders } from '@angular/common/http';
-import { Http, Headers } from '@angular/http';
-import 'rxjs/add/operator/toPromise';
+import { HttpClient, HttpHeaders } from '@angular/common/http';
 // import { saveAs } from 'file-saver/FileSaver';
-import { saveAs } from 'file-saver';
 import { ExcelService } from '../../services/excel/excel.service';
-import { constructor } from 'vfile-message';
-
 
 @Component({
   selector: 'tallyexport',
@@ -49,7 +42,7 @@ export class TallyexportComponent implements OnInit {
     public excelService: ExcelService,
     public csvService: CsvService,
     public accountService: AccountService,
-    private http: Http,
+    private http: HttpClient,
     public modalService: NgbModal) {
     this.accountService.fromdate = (this.accountService.fromdate) ? this.accountService.fromdate: this.startDate;
     this.accountService.todate = (this.accountService.todate)? this.accountService.todate: this.endDate;
