@@ -3,79 +3,79 @@ import { NbMenuItem } from '@nebular/theme';
 export const ADMIN_MENU_ITEMS = JSON.stringify([
   {
     title: 'Dashboard',
-    icon: 'fa fa-home',
+    icon: 'fa-home',
     link: '/pages/dashboard',
   },
   {
     title: 'Load-Intelligence',
-    icon: 'fa fa-home',
+    icon: 'fa-home',
     link: '/load-intelligence',
   },
   {
     title: 'Trip',
-    icon: 'fas fa-route',
+    icon: 'fa-route',
     children: [
       {
         title: 'Vehicle Status Change',
-        icon: 'fas fa-clipboard-check',
+        icon: 'fa-clipboard-check',
         link: '/admin/vehiclestatuschange',
       },
       {
         title: 'Trip Verification',
-        icon: 'fas fa-clipboard-check',
+        icon: 'fa-clipboard-check',
         link: '/admin/tripverification',
       },
       {
         title: 'VSC Ticket Audit',
-        icon: 'fas fa-ticket-alt',
+        icon: 'fa-ticket-alt',
         link: '/admin/vscticketaudit',
       },
 
       {
         title: 'Issue Alerts',
-        icon: 'fa fa-exclamation-triangle',
+        icon: 'fa-exclamation-triangle',
         link: '/admin/issue-alerts',
       },
       {
         title: 'Alert Related Issue',
-        icon: 'far fa-bell',
+        icon: 'fa-bell',
         link: '/admin/alert-related-issue',
         home: true,
       },
       {
         title: 'Trip Feedback Logs',
-        icon: 'fas fa-comment-dots',
+        icon: 'fa-comment-dots',
         link: '/admin/trip-status-feedback-logs',
         home: true,
       },
 
       {
         title: 'Vehicle GPS Trail',
-        icon: 'fas fa-receipt',
+        icon: 'fa-receipt',
         link: '/admin/vehicle-gps-trail',
         home: true,
       },
       {
         title: 'Vehicle  Distance',
-        icon: 'fas fa-route',
+        icon: 'fa-route',
         link: '/admin/vehicle-distance',
         home: true,
       },
       {
         title: 'Trip Site Rule',
-        icon: 'fas fa-file-signature',
+        icon: 'fa-file-signature',
         link: '/admin/trip-site-rule',
         home: true,
       },
       {
         title: 'Placement Site Rule',
-        icon: 'fas fa-file-signature',
+        icon: 'fa-file-signature',
         link: '/admin/placement-site-rule',
         home: true,
       },
       {
         title: 'Transport Agents',
-        icon: 'fas fa-address-card',
+        icon: 'fa-address-card',
         link: '/admin/transport-agents',
         home: true,
       },
@@ -84,13 +84,13 @@ export const ADMIN_MENU_ITEMS = JSON.stringify([
      
       {
         title: 'Trip Diagnosis',
-        icon: 'fa fa-arrows-alt',
+        icon: 'fa-arrows-alt',
         link: '/admin/trip-analysis',
         home: true,
       },
       {
         title: 'VSC Diagnosis',
-        icon: 'fa fa-arrows-alt',
+        icon: 'fa-arrows-alt',
         link: '/admin/vsc-diagnosis',
         home: true,
       },
@@ -98,38 +98,38 @@ export const ADMIN_MENU_ITEMS = JSON.stringify([
   },
   {
     title: 'Sites',
-    icon: 'fas fa-coins',
+    icon: 'fa-coins',
     home: true,
     children: [{
       title: 'Site Fencing',
-      icon: 'fas fa-coins',
+      icon: 'fa-coins',
       link: '/admin/site-fencing',
     },
     {
       title: 'Site Details',
-      icon: 'fa fa-sitemap',
+      icon: 'fa-sitemap',
       link: '/admin/site-details',
       home: true,
     },
     {
       title: 'Subsites',
-      icon: 'fa fa-arrows-alt',
+      icon: 'fa-arrows-alt',
       link: '/admin/sub-sites',
       home: true,
     },
     {
       title: 'Buffer Polyline ',
-      icon: 'fas fa-coins',
+      icon: 'fa-coins',
       link: '/admin/buffer-polyline',
     },
     {
       title: 'Transport Area Fencing',
-      icon: 'fas fa-coins',
+      icon: 'fa-coins',
       link: '/admin/transport-area',
     },
     {
       title: 'Locations/Alias',
-      icon: 'fas fa-map',
+      icon: 'fa-map',
       link: '/admin/locations',
     },
     ]
@@ -137,101 +137,101 @@ export const ADMIN_MENU_ITEMS = JSON.stringify([
 
   {
     title: 'Admin',
-    icon: 'fa fa-user',
+    icon: 'fa-user',
     home: true,
     children: [
       {
         title: 'Driver List',
-        icon: 'fa fa-male',
+        icon: 'fa-male',
         link: '/admin/driver-list',
         home: true,
       },
 
       {
         title: 'Vehicle Categories',
-        icon: 'fa fa-chart-bar',
+        icon: 'fa-chart-bar',
         link: '/admin/vehicles-view',
       },
 
       {
         title: 'Diagnostics',
-        icon: 'fa fa-stethoscope',
+        icon: 'fa-stethoscope',
         link: '/admin/diagnostics',
       },
 
       {
         title: 'User Role',
-        icon: 'fas fa-user-cog',
+        icon: 'fa-user-cog',
         link: '/admin/user-preferences',
       },
       {
         title: 'FO Activity Summary',
-        icon: 'fas fa-file-alt',
+        icon: 'fa-file-alt',
         link: '/admin/activity-summary',
         home: true,
       },
       {
         title: 'Ticket Subscribe',
-        icon: 'fas fa-info-circle',
+        icon: 'fa-info-circle',
         link: '/admin/ticket-subscribe',
         home: true,
       },
       {
         title: 'Cust Onboarding',
-        icon: 'fas fa-info-circle',
+        icon: 'fa-info-circle',
         link: '/admin/add-customer',
         home: true,
       },
       {
         title: 'Toll Transaction Summary',
-        icon: 'fas fa fa-road',
+        icon: 'fa-road',
         link: '/admin/toll-transaction-summary',
         home: true,
       },
       {
         title: 'Manual Toll Transaction Summary',
-        icon: 'fas fa fa-user',
+        icon: 'fa-user',
         link: '/admin/manual-toll-transaction-summary',
         home: true,
       },
       {
         title: 'Vehicle Wise Toll Transaction Summary',
-        icon: 'fas fa fa-car',
+        icon: 'fa-car',
         link: '/admin/vehiclewise-tolltransaction',
         home: true,
       },
       {
         title: 'Vouchers Summary',
-        icon: 'far fa-list-alt',
+        icon: 'fa-list-alt',
         link: '/admin/vouchers-summary',
         home: true,
       },
       {
         title: 'Challan Payment Request',
-        icon: 'far fa-list-alt',
+        icon: 'fa-list-alt',
         link: '/admin/challan-payment-request',
         home: true,
       },
       {
         title: 'Mv Gps Api Request',
-        icon: 'far fa-list-alt',
+        icon: 'fa-list-alt',
         link: '/admin/mv-gps-api-req',
         home: true,
       },
       {
         title: 'Captcha',
-        icon: 'far fa-list-alt',
+        icon: 'fa-list-alt',
         link: '/admin/captcha',
         home: true,
       },
       {
         title: 'Financial Account Summary',
-        icon: 'fa fa-pencil-square-o',
+        icon: 'fa-pencil-square-o',
         link: '/admin/financial-account-summary',
       },
       {
         title: 'Vehicle Document Mismatch Summary',
-        icon: 'fa fa-pencil-square-o',
+        icon: 'fa-pencil-square-o',
         link: '/admin/vehdocmismatch',
       },
 
@@ -239,90 +239,90 @@ export const ADMIN_MENU_ITEMS = JSON.stringify([
   },
   {
     title: 'Receipt & Invoice',
-    icon: 'fas fa-receipt',
+    icon: 'fa-receipt',
     children: [
       {
         title: 'Lorry Receipt Details',
-        icon: 'fas fa-receipt',
+        icon: 'fa-receipt',
         link: '/admin/lorry-receipt-details',
       },
 
 
       {
         title: 'LR Diagnostics',
-        icon: 'fas fa-diagnostics',
+        icon: 'fa-diagnostics',
         link: '/admin/lr-diagnostics',
       },
       {
         title: 'POD Dashboard',
-        icon: 'fa fa-bar-chart',
+        icon: 'fa-bar-chart',
         link: '/admin/pod-dashboard',
       },
       {
         title: 'Nearby Pods',
-        icon: 'fa fa-bar-chart',
+        icon: 'fa-bar-chart',
         link: '/admin/nearby-pods',
       },
     ]
   },
   {
     title: 'Customer Admin',
-    icon: 'fa fa-user',
+    icon: 'fa-user',
     children: [
       {
         title: 'Fo GPS Mapping',
-        icon: 'fas fa-map-marked-alt',
+        icon: 'fa-map-marked-alt',
         link: '/admin/gps-supplier-mapping',
         home: true,
       },
       {
         title: 'Vehicle GPS Detail',
-        icon: 'fas fa-map-pin',
+        icon: 'fa-map-pin',
         link: '/admin/vehicle-gps-detail',
       },
       {
         title: 'Group Managements',
-        icon: 'fa fa-users',
+        icon: 'fa-users',
         link: '/admin/group-managements',
       },
       {
         title: 'Ticket Properties',
-        icon: 'fas fa-ticket-alt',
+        icon: 'fa-ticket-alt',
         link: '/admin/ticket-properties',
       },
       {
         title: 'Escalation Matrix',
-        icon: 'fa fa-chart-bar',
+        icon: 'fa-chart-bar',
         link: '/admin/escalation-matrix',
       },
       // {
       //   title: 'Fo Vehicle Details',
-      //   icon: 'fas fa-info-circle',
+      //   icon: 'fa-info-circle',
       //   link: '/admin/vehicles-view',
       //   home: true,
       // },
       {
         title: 'Company Details',
-        icon: 'fas fa-info-circle',
+        icon: 'fa-info-circle',
         link: '/admin/company-details',
         home: true,
       },
       {
         title: 'Web Activity Summary',
-        icon: 'far fa-user',
+        icon: 'fa-user',
         link: '/admin/web-activity-summary',
         home: true,
       },
 
       {
         title: 'User Template',
-        icon: 'far fa-user',
+        icon: 'fa-user',
         link: '/admin/user-templates',
         home: true,
       },
       {
         title: 'Routes',
-        icon: 'fa fa-pencil-square-o',
+        icon: 'fa-pencil-square-o',
         link: '/admin/via-routes',
       },
     ]
@@ -331,16 +331,16 @@ export const ADMIN_MENU_ITEMS = JSON.stringify([
 
   {
     title: 'Document',
-    icon: 'fa fa-file-alt',
+    icon: 'fa-file-alt',
     children: [
       {
         title: 'Pending Details',
-        icon: 'fa fa-pencil-square-o',
+        icon: 'fa-pencil-square-o',
         link: '/admin/pending-documents',
       },
       {
         title: 'Pending vehicle Modal',
-        icon: 'fa fa-pencil-square-o',
+        icon: 'fa-pencil-square-o',
         link: '/admin/pending-vehicle',
       },
 
@@ -349,72 +349,72 @@ export const ADMIN_MENU_ITEMS = JSON.stringify([
 
   {
     title: 'Fuel',
-    icon: 'fas fa-gas-pump',
+    icon: 'fa-gas-pump',
     children: [
       {
         title: 'Fuel Fillings',
-        icon: 'fas fa-gas-pump',
+        icon: 'fa-gas-pump',
         link: '/admin/fuel-fillings',
         home: true,
       },
       {
         title: 'Fuel Mileage',
-        icon: 'fas fa-tachometer-alt',
+        icon: 'fa-tachometer-alt',
         link: '/admin/fuel-average-analysis',
         home: true,
       },
       {
         title: 'Consolidate Fuel Average',
-        icon: 'fa fa-signal',
+        icon: 'fa-signal',
         link: '/admin/consolidate-fuel-average',
         home: true,
       },
       {
         title: 'Fuel Diagnosis',
-        icon: 'fa fa-signal',
+        icon: 'fa-signal',
         link: '/admin/fuel-analysis',
         home: true,
       },
       {
         title: 'Remaining Fuel',
-        icon: 'fas fa-oil-can',
+        icon: 'fa-oil-can',
         link: '/admin/remaining-fuel',
         home: true,
       },
       {
         title: 'Fuel Rules',
-        icon: 'fa fa-signal',
+        icon: 'fa-signal',
         link: '/admin/fuel-rules',
         home: true,
       },
       {
         title: 'Fo Model Fuel Milage',
-        icon: 'fas fa-gas-pump',
+        icon: 'fa-gas-pump',
         link: '/admin/fo-fuel-average',
         home: true,
       },
 
       {
         title: 'Pump Station Area',
-        icon: 'fas fa-gas-pump',
+        icon: 'fa-gas-pump',
         link: '/admin/pump-station-area',
         home: true,
       },
       {
         title: 'Fuel Average Issues',
-        icon: 'fas fa-tachometer-alt',
+        icon: 'fa-tachometer-alt',
         link: '/admin/fuel-average-issues',
         home: true,
       },
       {
         title: 'Fo Fs Mapping',
-        icon: 'fa fa-signal',
+        icon: 'fa-signal',
         link: '/admin/fo-fs-mapping',
         home: true,
       },
       {
         title: 'Fuel Mileage With ODO',
-        icon: 'fas fa-gas-pump',
+        icon: 'fa-gas-pump',
         link: '/admin/fuel-mileage-with-odo',
         home: true
       },
@@ -422,21 +422,21 @@ export const ADMIN_MENU_ITEMS = JSON.stringify([
   },
   {
     title: 'Account',
-    icon: 'fa fa-money',
+    icon: 'fa-money',
     link: '/accounts/dashboard'
   },
   {
     title: 'Maintenance',
-    icon: 'fa fa-file-alt',
+    icon: 'fa-file-alt',
     children: [
       {
         title: 'Service Model',
-        icon: 'fa fa-pencil-square-o',
+        icon: 'fa-pencil-square-o',
         link: '/admin/view-modal-service',
       },
       {
         title: 'Sub Service Model',
-        icon: 'fa fa-pencil-square-o',
+        icon: 'fa-pencil-square-o',
         link: '/admin/view-sub-modal-service',
       },
 
@@ -444,11 +444,11 @@ export const ADMIN_MENU_ITEMS = JSON.stringify([
   },
   {
     title: 'Battery',
-    icon: 'fas fa-car-battery',
+    icon: 'fa-car-battery',
     children: [
       {
         title: 'Battery Models',
-        icon: 'fa fa-text-width',
+        icon: 'fa-text-width',
         link: '/admin/battery-modals',
         home: true,
       },
