@@ -7,27 +7,27 @@
 import { APP_BASE_HREF } from '@angular/common';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+
 import { NgModule } from '@angular/core';
 import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
 import { CoreModule } from './@core/core.module';
 import { ThemeModule } from './@theme/theme.module';
 import { NgbModule, NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';
-import { HttpModule } from '@angular/http';
 import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
 
 // BY On Demand add Module
-import { NgxQRCodeModule } from 'ngx-qrcode2';
-import { OwlDateTimeModule, OwlNativeDateTimeModule } from 'ng-pick-datetime';
+import { NgxQRCodeModule } from '@techiediaries/ngx-qrcode';
+import { OwlDateTimeModule, OwlNativeDateTimeModule } from 'ng-pick-datetime-ex';
 import { PdfViewerModule } from 'ng2-pdf-viewer';
 import { DirectiveModule } from './directives/directives.module';
 import { from } from 'rxjs';
 import { DragDropModule } from '@angular/cdk/drag-drop';
 import { ScrollingModule } from '@angular/cdk/scrolling';
-import { ImageViewerModule } from 'ng2-image-viewer';
 import { MatIconModule } from '@angular/material/icon';
-import { OWL_DATE_TIME_LOCALE } from 'ng-pick-datetime';
+// import { OWL_DATE_TIME_LOCALE } from 'ng-pick-datetime';
 import { ReactiveFormsModule } from '@angular/forms';
-import { MatDatepickerModule, MatNativeDateModule } from '@angular/material';
+import { MatNativeDateModule } from '@angular/material/core';
+import { MatDatepickerModule } from '@angular/material/datepicker';
 import { DeactivateGuardService } from './guards/route.guard';
 import { NgxPrintModule } from 'ngx-print';
 import { DateInputsModule } from '@progress/kendo-angular-dateinputs';
@@ -298,7 +298,9 @@ import { TicketDetailsComponent } from './modals/ticket-details/ticket-details.c
 import { ShowDataMapComponent } from './modals/generic-modals/show-data-map/show-data-map.component';
 import { IframeModalComponent } from './modals/iframe-modal/iframe-modal.component';
 import { UploadFileComponent } from './modals/generic-modals/upload-file/upload-file.component';
+import { NbEvaIconsModule } from '@nebular/eva-icons';
 import { FoSiteAliasComponent } from './modals/fo-site-alias/fo-site-alias.component';
+import { TripKmRepairViewComponent } from './modals/trip-km-repair-view/trip-km-repair-view.component';
 
 const COMMON_COMPONENT = [UnMergeStateComponent,
   PdfViewerComponent,
@@ -571,6 +573,7 @@ const COMMON_COMPONENT = [UnMergeStateComponent,
     VerifyfuturetripstateComponent,
     CheckloginandredirectComponent,
     CheckloginandredirectComponent,
+    TripKmRepairViewComponent,
 
   ],
   entryComponents: [
@@ -582,14 +585,12 @@ const COMMON_COMPONENT = [UnMergeStateComponent,
     BrowserAnimationsModule,
     HttpClientModule,
     AppRoutingModule,
-    HttpModule,
-    NgbModule.forRoot(),
+    NgbModule,
     ThemeModule.forRoot(),
     CoreModule.forRoot(),
     DirectiveModule,
     OwlDateTimeModule,
     OwlNativeDateTimeModule,
-    ImageViewerModule,
     MatIconModule,
     DragDropModule,
     ReactiveFormsModule,
@@ -601,7 +602,8 @@ const COMMON_COMPONENT = [UnMergeStateComponent,
     PdfViewerModule,
     ResizableModule,
     DateInputsModule,
-    ScrollingModule
+    ScrollingModule,
+    NbEvaIconsModule
   ],
 
   bootstrap: [AppComponent],
@@ -609,7 +611,7 @@ const COMMON_COMPONENT = [UnMergeStateComponent,
     DeactivateGuardService,
     NgbActiveModal,
     { provide: APP_BASE_HREF, useValue: '/' },
-    { provide: OWL_DATE_TIME_LOCALE, useValue: 'en' },
+    // { provide: OWL_DATE_TIME_LOCALE, useValue: 'en' },
     { provide: HTTP_INTERCEPTORS, useClass: HttpResponseHandlerService, multi: true },
     // {provide: OWL_DATE_TIME_FORMATS, useValue: 'MMMM YYYY'}
   ],
