@@ -5,7 +5,7 @@ import { ThemeModule } from '../@theme/theme.module';
 import { BatteryComponent } from './battery.component';
 import { BatteryInventoryComponent } from './battery-inventory/battery-inventory.component';
 import { DirectiveModule } from '../directives/directives.module';
-import { OwlDateTimeModule, OwlNativeDateTimeModule } from 'ng-pick-datetime';
+import { OwlDateTimeModule, OwlNativeDateTimeModule } from 'ng-pick-datetime-ex';
 import { BatteryModalsComponent } from './battery-modals/battery-modals.component';
 import { VehicleBatteryComponent } from './vehicle-battery/vehicle-battery.component';
 import { BatterySummaryComponent } from './battery-summary/battery-summary.component';
