@@ -3,41 +3,41 @@ import { NbMenuItem } from '@nebular/theme';
 export const BIDSYSTEM_MENU_ITEMS = JSON.stringify([
   {
     title: 'Home Dashboard',
-    icon: 'fa fa-home',
+    icon: 'fa-home',
     link: '/pages/dashboard',
     home: true,
 },
 {
     title: 'Dashboard',
-    icon: 'fa fa-credit-card',
+    icon: 'fa-credit-card',
     link: '/bid-system/dashboard',
     home: true,
 
 },
 {
   title: 'Order Boards',
-  icon: 'fa fa-credit-card',
+  icon: 'fa-credit-card',
   link: '/bid-system/order-boards',
   home: true,
 
 },
 {
   title: 'Open Orders',
-  icon: 'fa fa-credit-card',
+  icon: 'fa-credit-card',
   link: '/bid-system/open-orders',
   home: true,
 
 },
 {
   title: 'Orders-Won',
-  icon: 'fa fa-trophy',
+  icon: 'fa-trophy',
   link: '/bid-system/orders-won',
   home: true,
 
 },
 {
   title: 'Trip Management',
-  icon: 'fa fa-truck',
+  icon: 'fa-truck',
   link: '/bid-system/order-trip-management',
   home: true,
 
