@@ -2,13 +2,13 @@ import { NbMenuItem } from '@nebular/theme';
 export const CHALLAN_MENU_ITEMS = JSON.stringify([
     {
       title: 'Home Dashboard',
-      icon: 'fa fa-home',
+      icon: 'fa-home',
       link: '/pages/dashboard',
       home: true,
     },
     {
       title: 'Pending Challan',
-      icon: 'fa fa-home',
+      icon: 'fa-home',
       link: '/challan/pending-challan',
       home: true,
     },
