@@ -3,36 +3,36 @@ import { NbMenuItem } from '@nebular/theme';
 export const ADMIN_MENU_ITEMS: NbMenuItem[] = [
     {
         title: 'Home Dashboard',
-        icon: 'fa fa-home',
+        icon: 'fa-home',
         link: '/pages/dashboard',
         home: true,
     },
     {
         title: 'Documents Input ',
-        icon: 'fa fa-book',
+        icon: 'fa-book',
         link: '/documents/documentation-details',
         home: true,
     },
     {
         title: 'Pending Details',
-        icon: 'fa fa-pencil-square-o',
+        icon: 'fa-pencil-square-o',
         link: '/documents/pending-documents',
         home: true,
     },
 
     {
         title: 'Customer Dashboard',
-        icon: 'fa fa-user',
+        icon: 'fa-user',
         children: [
             {
                 title: 'Document Dashboard',
-                icon: 'fa fa-table',
+                icon: 'fa-table',
                 link: '/documents/documents-summary',
                 home: true,
             },
             {
                 title: 'Document Summary',
-                icon: 'fa fa-file-text',
+                icon: 'fa-file-text',
                 link: '/documents/dashboard',
                 home: true,
             },
@@ -47,7 +47,7 @@ export const CUSTOMER_MENU_ITEMS: NbMenuItem[] = [
 
     {
         title: 'Home Dashboard',
-        icon: 'fa fa-home',
+        icon: 'fa-home',
         link: '/pages/dashboard',
         home: true,
     },
@@ -58,17 +58,17 @@ export const CUSTOMER_MENU_ITEMS: NbMenuItem[] = [
         children: [
             {
                 title: 'Document Dashboard',
-                icon: 'fa fa-table',
+                icon: 'fa-table',
                 link: '/documents/documents-summary',
             },
             {
                 title: 'Document Summary',
-                icon: 'fa fa-file-text',
+                icon: 'fa-file-text',
                 link: '/documents/dashboard',
             },
             {
                 title: ' Vehicle Document History',
-                icon: 'fa fa-truck',
+                icon: 'fa-truck',
                 link: '/documents/documentation-details',
             }
 
