@@ -1,11 +1,10 @@
-import { Component, OnInit, Renderer } from '@angular/core';
+import { Component, OnInit, Renderer2 } from '@angular/core';
 import { ApiService } from '../../services/api.service';
 import { UserService } from '../../services/user.service';
 import { CommonService } from '../..//services/common.service';
 import { NgbActiveModal, NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { DomSanitizer } from '@angular/platform-browser';
 import { GenericSuggestionComponent } from '../generic-modals/generic-suggestion/generic-suggestion.component';
-import { asElementData } from '@angular/core/src/view';
 import { lstat } from 'fs';
 
 @Component({
@@ -35,7 +34,7 @@ export class SaveUserTemplateComponent implements OnInit {
     public user: UserService,
     public activeModal: NgbActiveModal,
     private modalService: NgbModal,
-    public renderer: Renderer,
+    public renderer: Renderer2,
     private sanitizer: DomSanitizer
 
   ) {
