@@ -1,4 +1,4 @@
-import { Component, OnInit, Renderer } from '@angular/core';
+import { Component, OnInit, Renderer2 } from '@angular/core';
 import { ApiService } from '../../services/api.service';
 import { CommonService } from '../../services/common.service';
 import { UserService } from '../../services/user.service';
@@ -49,7 +49,7 @@ export class LorryRecciptsComponent implements OnInit {
     public user: UserService,
     public route: ActivatedRoute,
     private modalService: NgbModal,
-    public renderer: Renderer) {
+    public renderer: Renderer2) {
     this.getLorryReceipts();
     this.common.refresh = this.refresh.bind(this);
 
