@@ -6,6 +6,7 @@ import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { DatePipe } from '@angular/common';
 import { RemarkModalComponent } from '../../modals/remark-modal/remark-modal.component';
 import { RouteMapperComponent } from '../../modals/route-mapper/route-mapper.component';
+import { TripKmRepairViewComponent } from '../../modals/trip-km-repair-view/trip-km-repair-view.component';
 
 @Component({
   selector: 'trips',
@@ -156,9 +157,8 @@ export class TripsComponent implements OnInit {
         } else {
           this.valobj[this.headings[j]] = { value: this.vehicleTrips[i][this.headings[j]], class: 'black', action: '' };
         }
-        this.valobj['km'] = {
-          value: this.vehicleTrips[i]['_km'], isHTML: true, action: null,
-        }
+        this.valobj['km'] = { value: this.vehicleTrips[i]['_km'], class: 'blue',action:this.openTripKmRepair.bind(this, this.vehicleTrips[i]) };
+        
         this.valobj['googlekm'] = {
           value: this.vehicleTrips[i]['_googlekm'], isHTML: true, action: null,
         }
@@ -185,6 +185,23 @@ export class TripsComponent implements OnInit {
 
     console.log('Columns:', columns);
     return columns;
+  }
+
+  openTripKmRepair(tripInfo){
+    if(!tripInfo['_km']){
+      this.common.showError('No Data');
+      return;
+    }
+    let tripData = {
+      tripId : tripInfo['Trip Id']
+    };
+    this.common.params = tripData;
+    console.log("tripData", this.common.params);
+
+    const activeModal = this.modalService.open(TripKmRepairViewComponent, { size: 'lg', container: 'nb-layout' });
+    activeModal.result.then(data => {
+      // this.getVehicleTrips();
+    });
   }
 
   actionIcons(trip) {
