@@ -4,8 +4,7 @@ import { FuelAverageAnalysisComponent } from './pages/fuel-average-analysis/fuel
 import { ThemeModule } from './@theme/theme.module';
 import { DashboardModule } from './partner/dashboard/dashboard.module';
 import { DirectiveModule } from './directives/directives.module';
-import { OwlDateTimeModule, OwlNativeDateTimeModule } from 'ng-pick-datetime';
-import { ImageViewerModule } from 'ng2-image-viewer';
+import { OwlDateTimeModule, OwlNativeDateTimeModule } from 'ng-pick-datetime-ex';
 import { RemainingFuelComponent } from './admin/remaining-fuel/remaining-fuel.component';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 import { DocumentationDetailsComponent } from './documents/documentation-details/documentation-details.component';
@@ -63,7 +62,6 @@ const PAGES_COMPONENTS = [
         DirectiveModule,
         OwlDateTimeModule,
         OwlNativeDateTimeModule,
-        ImageViewerModule,
         ChartModule,
     ],
     exports: [...PAGES_COMPONENTS],
