@@ -3,61 +3,61 @@ import { NbMenuItem } from '@nebular/theme';
 export const TYRES_MENU_ITEMS = JSON.stringify([
     {
         title: 'Home Dashboard',
-        icon: 'fa fa-home',
+        icon: 'fa-home',
         link: '/pages/dashboard',
         home: true,
     },
     // {
     //     title: 'Inventory',
-    //     icon: 'fa fa-recycle',
+    //     icon: 'fa-recycle',
     //     link: '/tyres/inventory',
     //     home: true,
     // },
     {
         title: ' Vehicle Tyre Summary',
-        icon: 'fa fa-text-width',
+        icon: 'fa-text-width',
         link: '/tyres/tyre-summary',
         home: true,
     },
     {
         title: 'Vehicle Tyres',
-        icon: 'fa fa-pencil-square-o',
+        icon: 'fa-pencil-square-o',
         link: '/tyres/vehicle-tyres',
         home: true,
     },
     {
         title: 'Inputs',
-        icon: 'fa fa-pencil-square-o',
+        icon: 'fa-pencil-square-o',
         link: '/tyres/inputs',
         home: true,
     },
     // {
     //     title: 'Health Checkup',
-    //     icon: 'fa fa-medkit',
+    //     icon: 'fa-medkit',
     //     link: '/tyres/tyre-health-check-up',
     //     home: true,
     // },
     {
         title: 'Add Trolly',
-        icon: 'fa fa-truck',
+        icon: 'fa-truck',
         link: '/tyres/add-trolly',
         home: true,
     },
     {
         title: 'Vehicle Trolly Mapping',
-        icon: 'fa fa-street-view',
+        icon: 'fa-street-view',
         link: '/tyres/vehicle-trolly-mapping',
         home: true,
     },
     {
         title: 'Tyre Modals',
-        icon: 'fa fa-text-width',
+        icon: 'fa-text-width',
         link: '/tyres/tyre-modals',
         home: true,
     },
     {
         title: 'Tyre Summary',
-        icon: 'fa fa-text-width',
+        icon: 'fa-text-width',
         link: '/tyres/tyre-summary-details',
         home: true,
     },
