@@ -13,7 +13,7 @@ import { AddTrollyComponent } from './add-trolly/add-trolly.component';
 import { TyreModalsComponent } from './tyre-modals/tyre-modals.component';
 
 import { VehicleTrollyMappingComponent } from './vehicle-trolly-mapping/vehicle-trolly-mapping.component';
-import { OwlDateTimeModule, OwlNativeDateTimeModule } from 'ng-pick-datetime';
+import { OwlDateTimeModule, OwlNativeDateTimeModule } from 'ng-pick-datetime-ex';
 import { VehicleSearchComponent } from '../modals/vehicle-search/vehicle-search.component';
 import { VehicleTyresComponent } from './vehicle-tyres/vehicle-tyres.component';
 import { TyreSummaryComponent } from './tyre-summary/tyre-summary.component';
