@@ -3,25 +3,25 @@ import { NbMenuItem } from '@nebular/theme';
 export const MAINTENANCE_MENU_ITEMS = JSON.stringify([
     {
         title: 'Home Dashboard',
-        icon: 'fa fa-home',
+        icon: 'fa-home',
         link: '/pages/dashboard',
         home: true,
     },
     {
         title: 'Add Maintenance',
-        icon: 'fas fa-wrench',
+        icon: 'fa-wrench',
         link: '/vehicle-maintenance/add-vehicle-maintenance',
         home: true,
     },
     // {
     //     title: 'Maintenance Dashboard',
-    //     icon: 'fa fa-home',
+    //     icon: 'fa-home',
     //     link: '/vehicle-maintenance/maintenanace-dashboard',
     //     home: true,
     // },
     {
         title: 'Maintenance Summary',
-        icon: 'fas fa-cogs',
+        icon: 'fa-cogs',
         link: '/vehicle-maintenance/maintenance-summary',
         home: true,
     },
@@ -32,7 +32,7 @@ export const MAINTENANCE_MENU_ITEMS = JSON.stringify([
 
 //     {
 //         title: 'Home Dashboard',
-//         icon: 'fa fa-home',
+//         icon: 'fa-home',
 //         link: '/vehicle-maintenance/dashboard',
 //         home: true,
 //     },
