@@ -3,111 +3,111 @@ import { NbMenuItem } from '@nebular/theme';
 export const WALLE8_MENU_ITEMS = JSON.stringify([
     {
         title: 'Home Dashboard',
-        icon: 'fa fa-home',
+        icon: 'fa-home',
         link: '/pages/dashboard',
         home: true,
     },
     {
         title: 'Card Balance',
-        icon: 'fa fa-credit-card',
+        icon: 'fa-credit-card',
         link: '/walle8/card-balance',
         home: true,
 
     },
     {
         title: 'Card Mapping',
-        icon: 'fa fa-credit-card',
+        icon: 'fa-credit-card',
         link: '/walle8/card-mapping',
         home: true,
     },
     {
         title: 'Card Usage',
-        icon: 'fa fa-credit-card',
+        icon: 'fa-credit-card',
         link: '/walle8/card-usage',
         home: true,
     },
     {
         title: 'Payment Made',
-        icon: 'fa fa-credit-card',
+        icon: 'fa-credit-card',
         link: '/walle8/payments-made',
         home: true,
     },
     {
         title: 'Latest Recharge',
-        icon: 'fa fa-credit-card',
+        icon: 'fa-credit-card',
         link: '/walle8/latest-recharge',
         home: true,
     },
     {
         title: 'Toll Usage Summary',
-        icon: 'fa fa-credit-card',
+        icon: 'fa-credit-card',
         link: '/walle8/toll-usage-summary',
         home: true,
     },
     {
         title: 'Toll Discount',
-        icon: 'fa fa-credit-card',
+        icon: 'fa-credit-card',
         link: '/walle8/toll-discount',
         home: true,
     },
     {
         title: 'Toll Usage',
-        icon: 'fa fa-credit-card',
+        icon: 'fa-credit-card',
         link: '/walle8/toll-usage',
         home: true,
     },
     {
         title: 'Toll Settled Request',
-        icon: 'fa fa-credit-card',
+        icon: 'fa-credit-card',
         link: '/walle8/toll-setteled-request',
         home: true,
 
     },
     {
         title: 'Current Tag Balance',
-        icon: 'fa fa-credit-card',
+        icon: 'fa-credit-card',
         link: '/walle8/current-tag-balance',
         home: true,
 
     },
     {
         title: 'Double Toll Report',
-        icon: 'fa fa-credit-card',
+        icon: 'fa-credit-card',
         link: '/walle8/double-toll-report',
         home: true,
 
     },
     {
         title: 'Financial Toll Summary',
-        icon: 'fa fa-credit-card',
+        icon: 'fa-credit-card',
         link: '/walle8/financial-toll-summary',
         home: true,
 
     },
     {
         title: 'Financial Toll Summary(Add Time)',
-        icon: 'fa fa-credit-card',
+        icon: 'fa-credit-card',
         link: '/walle8/financial-toll-summary-addtime',
         home: true,
 
     },
     {
         title: 'Financial Main Summary',
-        icon: 'fa fa-credit-card',
+        icon: 'fa-credit-card',
         link: '/walle8/financial-main-summary',
         home: true,
 
     },
     {
         title: 'Toll Analytics',
-        icon: 'fa fa-credit-card',
+        icon: 'fa-credit-card',
         link: '/walle8/toll-analatics',
         home: true,
 
     },
     {
         title: 'Other Usage',
-        icon: 'fa fa-credit-card',
+        icon: 'fa-credit-card',
         link: '/walle8/other-usage',
         home: true,
 
